+++ src/tess/src/syntax.c
@@ -36,7 +36,7 @@ void syntax_checker_destroy(syntax_checker **self) {
 
 // -- syntax_checker operation --
 
-int syntax_checker_run(syntax_checker *self, ast_node **nodes, size_t count) {
+int syntax_checker_run(syntax_checker *self, ast_node **nodes, u32 count) {
 
     int res = 0;
     if ((res = syntax_rename_variables(self->alloc, nodes, count))) return res;
